Working version 1

+++ src/com/skilldistillery/jets/JetImpl.java
@@ -2,7 +2,7 @@ package com.skilldistillery.jets;
 
 public class JetImpl extends Jet {
 
-	public JetImpl(String model, int speed, int range, long price) {
+	public JetImpl(String model, double speed, int range, long price) {
 		super(model, speed, range, price);
 	}
